+++ examples/errors.py
@@ -2,17 +2,10 @@ from dataclasses import dataclass
 
 from koda import Err, Maybe
 
-from koda_validate.processors import strip
-from koda_validate.string import StringValidator, not_blank
-from koda_validate.validators.dicts import dict_validator
-from koda_validate.validators.validators import (
-    Choices,
-    IntValidator,
-    Min,
-    MinLength,
-    key,
-    maybe_key,
-)
+from koda_validate.dictionary import dict_validator, key, maybe_key
+from koda_validate.generic import Choices, Min
+from koda_validate.integer import IntValidator
+from koda_validate.string import MinLength, StringValidator, not_blank, strip
 
 # wrong type
 assert StringValidator()(None) == Err(["expected a string"])
