+++ examples/is_close_predicate.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 
 from koda import Err, Ok
 
+from koda_validate.float import FloatValidator
 from koda_validate.typedefs import JSONValue, Predicate
-from koda_validate.validators.validators import FloatValidator
 
 
 @dataclass
