+++ koda_validate/processors.py
@@ -1,25 +1,1 @@
 from koda_validate.typedefs import Processor
-
-
-class Strip(Processor[str]):
-    def __call__(self, val: str) -> str:
-        return val.strip()
-
-
-strip = Strip()
-
-
-class UpperCase(Processor[str]):
-    def __call__(self, val: str) -> str:
-        return val.upper()
-
-
-upper_case = UpperCase()
-
-
-class LowerCase(Processor[str]):
-    def __call__(self, val: str) -> str:
-        return val.lower()
-
-
-lower_case = LowerCase()
