+++ pyproject.toml
@@ -55,7 +55,7 @@ addopts ="""
 
 [tool.coverage.run]
 omit = [
-    "koda_validate/dict.py",
+    "koda_validate/dictionary.py",
     "koda_validate/validate_and_map.py"
 ]
