--- tests/test_processors.py
@@ -1,14 +1,0 @@
-from koda_validate.processors import lower_case, strip, upper_case
-
-
-def test_strip() -> None:
-    assert strip(" x ") == "x"
-    assert strip("ok") == "ok"
-
-
-def test_upper_case() -> None:
-    assert upper_case("AbCdE") == "ABCDE"
-
-
-def test_lower_case() -> None:
-    assert lower_case("ZyXwV") == "zyxwv"
